+++ project/src/components/property-goods-list/property-goods-list.tsx
@@ -1,5 +1,3 @@
-import PropertyGood from './property-good';
-
 type PropertyGoodsProps = {
   goods: string[];
 }
@@ -7,7 +5,7 @@ type PropertyGoodsProps = {
 function PropertyGoodsList({goods}:PropertyGoodsProps):JSX.Element{
   return(
     <ul className="property__inside-list">
-      {goods.map((item)=>(<PropertyGood key={item} good={item}/>))}
+      {goods.map((item)=>(<li key={item} className="property__inside-item">{item}</li>))}
     </ul>
   );
 }
